Document v0.5.0 test results and refine type names
Adds a detailed status report and test results for the v0.5.0 milestone, highlighting an 88% success rate and identifying remaining tasks.

Standardizes the display of arbitrary-precision integers as "int" to provide a more consistent user-facing type system.

Updates unit tests and documentation to reflect recent changes in type definitions and lexer tokens.

+++ src/ast/types.rs
@@ -333,7 +333,7 @@ impl std::fmt::Display for Type {
             Type::Bool => write!(f, "bool"),
             Type::Str => write!(f, "str"),
             Type::Bytes => write!(f, "bytes"),
-            Type::BigInt => write!(f, "bigint"),
+            Type::BigInt => write!(f, "int"),
             Type::None => write!(f, "None"),
             Type::Int => write!(f, "int"),
             Type::List(t) => write!(f, "[{}]", t),
